Start making use of branches to stay organized
It was discussed in the first meeting (Friday, April 9, 2021) that we
would work on Task 1 together this weekend and try to complete by the
following Monday. It was also proposed that at some point (after
having) completed a task or two, that one person may split off and
begin working on the bonus (task 5) by themselves. If we do this we
will want to make sure that we stay organized, branches by name of task
is a good way to do this.

A pull request will be created for each task with task list boxes (the
boxes are empty if the task is not done, they are checked if it is
completed). Once all boxes are checked, and the team agrees that the
task is complete, the branch will be merged into the master branch.