Add in modified csv data
The rows at the bottom of the csv file have been removed. As Dr. Thomas
mentioned in class, any rows of data that are complete garbage can be
removed by hand. The information was helpful for determining what was
in the dataset, but for processing purposes with spark it will be
easier if the data is not in the file.