Fix the splitting dataframe procedure
There was an error in splitting the complete dataframe into two
dataframes: a dataframe containing only good values and dataframe
containing only bad. This problem has been corrected.

The way to check that this is done properly is to sum the number of
rows of the good and bad dataframe and see if it is equivalent to the
original number of rows.