Add example to verify the normalization process
It is believed that for cosine similarity to work that the data must
first be normalized. After the most similar row is found the data may
then need to be denormalized to find its true value. In the
normalization process the data is jumbled up, so there is no real way
to map the values to each other. That is unless the datatframe contains
both the un-normalized data and the normalized data.