Modularize normalization and remove denormalization
The normalization functions have been seperated so that they can be
better used in this process.

The denormalization is not necessary because an indexing scheme will be
used. The normalization will be placed into the same dataframe as where
the data exists. The original values will be unchanged. This is a more
straightforward process than was previously planned.